move endmember_extraction files into workspace or demo
