+++ hsi_toolkit/__init__.py
@@ -1,1 +1,2 @@
 from anomaly_detectors import *
+from endmember_extraction import *
